requirements.txt
fastapi
uvicorn[standard]
pyodbc
openai
PyYAML
python-dotenv
pandas
sqlglot
streamlit
requests